restructured network into rpc and sever
