+++ src/main.go
@@ -2,17 +2,19 @@ package main
 
 import (
 	"D7024E/log"
+	"D7024E/network/server"
 	"D7024E/node/bucket"
 	"D7024E/node/contact"
 	"D7024E/node/id"
 	"D7024E/node/stored"
+	"net"
 )
 
 func main() {
-	// ip := net.IPv4(127, 0, 0, 1)
-	// port := 4001
-	// go network.UDPListener(ip, port)
-	// network.UDPSender(ip, port, "this is the message")
+	ip := net.IPv4(127, 0, 0, 1)
+	port := 4001
+	go server.UDPListener(ip, port)
+	server.UDPSender(ip, port, "this is the message")
 
 	rt := bucket.GetInstance()
 	rt.SetMe(contact.Contact{ID: id.NewRandomKademliaID(), Address: "this is address"})
