+++ src/network/serverREST/controller/object/getObject.go
