+++ src/network/serverREST/controller/object/getObjects.go
